requirements.txt
python-dotenv
aiohttp
websockets
curl_cffi
httpx
kickpython